requirements.txt
docling
langchain_community 
openai 
markdown2 
python-dotenv
